This adds support for generic values in the tree.
Keys still are limited to strings.

This was born from an attempt at making the structure fully generic, but I couldn't find a nice way to make the keys generic without sacrificing the ergonomics of the `string` case.

Backward compatibility is kept for simple tree usage, but will require explicit types e.g. for `*Iterator` variables (as can be seen in `TestWalk`).

The change currently has a slight performance penalty on the `Get` case, and a slight performance improvement on the `Put` case:
```
name              old time/op    new time/op    delta
Get/Words-8         23.5ms ± 3%    25.7ms ± 4%  +9.50%  (p=0.008 n=5+5)
Get/Web2a-8         7.18ms ± 7%    7.99ms ±12%    ~     (p=0.056 n=5+5)
Put/Words-8         69.4ms ± 4%    68.0ms ± 2%    ~     (p=0.222 n=5+5)
Put/Web2a-8         22.5ms ± 2%    22.4ms ± 2%    ~     (p=0.548 n=5+5)
Walk/Words-8        7.41ms ± 7%    7.52ms ± 8%    ~     (p=0.841 n=5+5)
Walk/Web2a-8        1.43ms ±13%    1.54ms ± 9%    ~     (p=0.151 n=5+5)
WalkPath/Words-8    25.5ms ± 3%    27.0ms ± 5%  +5.99%  (p=0.008 n=5+5)
WalkPath/Web2a-8    7.59ms ± 4%    7.93ms ± 5%    ~     (p=0.151 n=5+5)

name              old alloc/op   new alloc/op   delta
Get/Words-8          0.00B          0.00B         ~     (all equal)
Get/Web2a-8          0.00B          0.00B         ~     (all equal)
Put/Words-8         36.5MB ± 0%    36.5MB ± 0%    ~     (p=0.171 n=4+4)
Put/Web2a-8         12.5MB ± 0%    12.5MB ± 0%    ~     (p=0.603 n=5+5)
Walk/Words-8         0.00B          0.00B         ~     (all equal)
Walk/Web2a-8         0.00B          0.00B         ~     (all equal)
WalkPath/Words-8     0.00B          0.00B         ~     (all equal)
WalkPath/Web2a-8     0.00B          0.00B         ~     (all equal)

name              old allocs/op  new allocs/op  delta
Get/Words-8           0.00           0.00         ~     (all equal)
Get/Web2a-8           0.00           0.00         ~     (all equal)
Put/Words-8          1.07M ± 0%     1.07M ± 0%    ~     (all equal)
Put/Web2a-8           352k ± 0%      352k ± 0%    ~     (all equal)
Walk/Words-8          0.00           0.00         ~     (all equal)
Walk/Web2a-8          0.00           0.00         ~     (all equal)
WalkPath/Words-8      0.00           0.00         ~     (all equal)
WalkPath/Web2a-8      0.00           0.00         ~     (all equal)
```

WDYT?