@gammazero 2023-02-24
I specifically do not want to do this because it forces a tree to store a specific type of value. While that is the typical use-case, there are also times when the same tree is used to store different types.
```go
rt := new(Tree)                                                                                                                                   
rt.Put("company", "Weyland-Yutani")                                                                                                                                     
rt.Put("order", 939)
```