@costela 2023-02-24
It's of course totally your decision, but wouldn't `Tree[any]` still satisfy the mentioned use-case? :thinking: